face_detection/requirements.txt
numpy
matplotlib
opencv-python
pandas
Pillow
torch
torchvision
